MADP/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.5.1
networkx==3.5
numpy==2.3.0
nvidia-ml-py==12.575.51
nvitop==1.5.1
packaging==25.0
pandas==2.3.0
pillow==11.0.0
propcache==0.3.2
psutil==7.0.0
pyglet==1.5.31
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rware==2.0.0
setuptools==78.1.1
six==1.17.0
sympy==1.13.1
torch==2.6.0+cu126
torch-geometric==2.6.1
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
windows-curses==2.4.1
yarl==1.20.1
